Refactor: Update Firebase timestamp handling and path configuration
This commit introduces several changes related to Firebase data handling:

- **Firebase Path Configuration:**
    - Introduced `FirebaseConstants.kt` to define `SMS_NODES_PATH`.
    - `SMS_NODES_PATH` is now configured via `buildConfigField` in `app/build.gradle.kts`:
        - `"sms_by_date_debug"` for debug builds.
        - `"sms_by_date"` for release builds.
    - `TransactionsInFBActivity.kt` and `TransactionProcessor.kt` now use `FirebaseConstants.SMS_NODES_PATH` for constructing Firebase paths.
- **Timestamp Handling:**
    - Changed the `date` field in `TransactionInfo` from `Long?` to `Int` (representing Unix epoch seconds).
    - Updated `FirebaseSyncState.kt` to use `Int` for `smsLastMsgTimestamp` and related methods.
    - `TransactionsInFBActivity.kt`:
        - Sample data generation now uses `(calendar.timeInMillis / 1000).toInt()` for the `date` field.
        - `parseTransactionNode` now expects the `date` field from Firebase to be an `Int`. The complex logic for handling string dates or mixed types for `date` and `strDateInMessage` has been largely removed or simplified, prioritizing the `Int` timestamp.
    - `TransactionsAdapterFB.kt`:
        - The `date` in `AdapterItemFB.DateHeaderItem` is now an `Int`.
        - Date and time formatting now uses `(timestamp * 1000L).formatDateOrTime(...)` for display, leveraging `fossify.commons.extensions`.
        - Sorting of transactions now uses the `Int` date.
    - `TransactionProcessor.kt`:
        - `parseMessage` now takes a `Message` object as input (instead of just the body `String`) to access `message.date` (which is an `Int`) for populating `TransactionInfo.date`.
    - `MainActivity.kt`:
        - `syncNewSmsToFirebase` now uses `message.date` directly when creating `TransactionInfo`.
        - Removed unused `convertToTimestamp` and `formatDateForFirebasePath` methods.
    - `SmsReceiver.kt`: Passes the full `message` object to `TransactionProcessor.parseMessage`.
- **UI and Adapter Changes:**
    - `TransactionsAdapterFB` constructor now takes a `SimpleActivity` instance, likely for context-dependent operations like string formatting.
    - `TransactionsInFBActivity` now extends `SimpleActivity`.
- **Code Cleanup:**
    - Commented out the `loadTransactions` logic in `TransactionActivity.kt`, effectively disabling SMS loading from the device in this activity.
    - Removed the `addDummyTransactionToFB` method from `TransactionsInFBActivity.kt`.

+++ app/src/main/kotlin/org/fossify/messages/helpers/FirebaseSyncState.kt
@@ -21,13 +21,13 @@ object FirebaseSyncState {
     private val scope = CoroutineScope(Dispatchers.IO)
 
     @Volatile
-    private var smsLastMsgTimestamp: Long = 0L
+    private var smsLastMsgTimestamp: Int = 0
     @Volatile
     private var initialized: Boolean = false
     @Volatile
     private var initializing: Boolean = false
 
-    fun initialize(callback: ((success: Boolean, timestamp: Long) -> Unit)? = null) {
+    fun initialize(callback: ((success: Boolean, timestamp: Int) -> Unit)? = null) {
         if (initialized || initializing) {
             callback?.invoke(initialized, smsLastMsgTimestamp)
             return
@@ -38,7 +38,7 @@ object FirebaseSyncState {
         database.child(FB_GLOBALS_PATH).child(FB_SMS_LAST_MSG_TIMESTAMP_KEY)
             .addListenerForSingleValueEvent(object : ValueEventListener {
                 override fun onDataChange(snapshot: DataSnapshot) {
-                    val timestamp = snapshot.getValue(Long::class.java) ?: 0L
+                    val timestamp = snapshot.getValue(Int::class.java) ?: 0
                     smsLastMsgTimestamp = timestamp
                     initialized = true
                     initializing = false
@@ -49,7 +49,7 @@ object FirebaseSyncState {
                 override fun onCancelled(error: DatabaseError) {
                     TODO("Not yet implemented")
                     Log.e(TAG, "Error initializing smsLastMsgTimestamp: ${error.message}")
-                    smsLastMsgTimestamp = 0L // Default to 0 on error
+                    smsLastMsgTimestamp = 0 // Default to 0 on error
                     initialized = false // Consider if an error means it's not truly initialized
                     initializing = false
                     callback?.invoke(false, smsLastMsgTimestamp)
@@ -57,14 +57,14 @@ object FirebaseSyncState {
             })
     }
 
-    fun getLastMsgTimestamp(): Long {
+    fun getLastMsgTimestamp(): Int {
         if (!initialized && !initializing) {
             Log.w(TAG, "Accessed getLastMsgTimestamp before initialization. Consider calling initialize() first.")
         }
         return smsLastMsgTimestamp
     }
 
-    fun updateLastMsgTimestampInFirebase(newTimestamp: Long) {
+    fun updateLastMsgTimestampInFirebase(newTimestamp: Int) {
         if (!initialized && !initializing) {
             // If not initialized, it means we don't know the baseline from FB.
             // Updating FB without this baseline could lead to inconsistencies.
@@ -90,7 +90,7 @@ object FirebaseSyncState {
     }
 
     // Call this if you want to ensure it's initialized before proceeding using coroutines
-    suspend fun awaitInitialization(): Long {
+    suspend fun awaitInitialization(): Int {
         if (initialized) return smsLastMsgTimestamp
         if (initializing) {
             // A simple way to wait for ongoing initialization if called from a coroutine.
